move x tests to single file
