--- x/jwt.test.ts
@@ -1,20 +1,0 @@
-// Copyright 2023 Samuel Kopp. All rights reserved. Apache-2.0 license.
-import {
-  assertEquals,
-  assertInstanceOf,
-} from 'https://deno.land/std@0.194.0/testing/asserts.ts'
-import { createKey, importKey, jwt } from './mod.ts'
-
-Deno.test('x/jwt', async (t) => {
-  const key = await createKey()
-  const cryptoKey = await importKey(key)
-
-  assertInstanceOf(cryptoKey, CryptoKey)
-
-  await t.step('sign/verify', async () => {
-    const token = await jwt.sign({ example: 'object' }, key)
-
-    assertEquals(await jwt.verify(token, key) !== undefined, true)
-    assertEquals(await jwt.verify(token, cryptoKey) !== undefined, true)
-  })
-})
